use go-logging instead of logrus, remove CallHandlerForMatch
logrus has an extensive feature set that is not particulary useful for the engine. The go-logging package has a simpler API and setup, and is a more lightweight solution.

The callHandlerForMatch function is removed for performance reasons: it performed significantly worse than a switch with type assertion (~60 ns/op vs ~2.7).